crewai_Agentic_AI/pyproject.toml
[project]
name = "crewai-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.1",
    "crewai>=0.159.0",
    "crewai-tools[mcp]>=0.62.3",
    "docx>=0.2.4",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "markdown>=3.8.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "sse-starlette>=3.0.2",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
    "xhtml2pdf>=0.2.17",
]
